circle cavity on sphere
Paramerters are defined to simulate light transmit in a circle cavity on sphere,where light would  propagate along the geodesic line. For sphere the geodesic line is big circle center in sphere center. You may set sphere radius, deformation parameter, the first light base point and direction. The project could give the picture of the cavity and light trajectory of every transmission. The PSOS result is stored in paramater q, which could be displayed by listplay function. In the end of the project, there are some results the author has calculated, which readers could use as  references.